+++ particle.js
@@ -14,7 +14,7 @@ class Particle {
 		this.vel.limit(this.maxSpeed);
 		this.pos.add(this.vel);
 		this.acc.mult(0);
- 
+		
 		this.edges();
 	}